requirements.txt
streamlit
firebase-admin
google-cloud-firestore
pandas